% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap.mfa}
\alias{bootstrap.mfa}
\title{bootstrap.mfa}
\usage{
bootstrap.mfa(X)
}
\arguments{
\item{X:}{a mfa object}
}
\value{
result:...
}
\description{
a brief description
}
\examples{
\dontrun{
an example}
}

