% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{contribution}
\alias{contribution}
\title{contribution}
\usage{
contribution(x, type = "table")
}
\arguments{
\item{x}{a mfa object}

\item{type}{a character-"observation","variable" or "table"}
}
\value{
The function will reutrn a data frame.

Each componets of the data frame represeting the proportion of each element's
        contribution to the certain dimension
}
\description{
contribution is the function of MFA package.
             The function aims to analyze the contributions of important elements to dimensions.
}

