% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa}
\alias{mfa}
\title{mfa}
\usage{
mfa(data, sets, ncomps = NULL, center = TRUE, scale = TRUE)
}
\arguments{
\item{data:}{data that need to be analysed}
}
\value{
result:...
}
\description{
a brief description
}
\examples{
\dontrun{
an example}
}

