% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{plot.boostrap}
\alias{plot.boostrap}
\title{plot.boostrap}
\usage{
\method{plot}{boostrap}(X, L)
}
\arguments{
\item{X:}{a mfa object}

\item{L:}{number of bootstrap samples are computed}
}
\value{
figure 7 & 8 in paper
}
\description{
plot figure 7 & 8 in paper, use AFTER running bootsrap.mfa()
}
\examples{
\dontrun{
an example}
}

