% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mfa}
\alias{plot.mfa}
\title{plot.mfa}
\usage{
\method{plot}{mfa}(x, dim_plot = c(1, 2), cex = 0.8, mfrow_input = c(2,
  5), ...)
}
\arguments{
\item{x}{data set(class should be mfa)}

\item{dim_plot}{vector of 2 indicating the 2 dimensions to plot partial factor scores with
(integer between 1 to 11).}

\item{cex}{positive number indicating the size of texts in graphs.}

\item{mfrow_input}{vector of 2 indicating how many graphs in each row and col.}

\item{...}{other arguments}
}
\value{
4 graphs
}
\description{
plot 4 graphs containing common factor scores, partial factor scores, and loadings
}
\examples{
\dontrun{
plot.mfa(review, dim_plot = c(1,2), cex = 0.8, mfrow_input = c(2,5))}
}

