% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mfa}
\alias{print.mfa}
\title{print.mfa}
\usage{
\method{print}{mfa}(x, n_assessor = 1, ...)
}
\arguments{
\item{x}{data set(class should be mfa)}

\item{n_assessor}{integer between 1 to 10 indicating the number of the chosen assessor.}

\item{...}{other arguments}
}
\value{
4 tables
}
\description{
print 4 tables containing eigen values, common factor scores, partial facotr scores, and loadings.
}
\examples{
\dontrun{
print.mfa(review, n_assessor = 1)}
}

