% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{RV_table}
\alias{RV_table}
\title{RV_table}
\usage{
RV_table(dataset, sets)
}
\arguments{
\item{dataset}{a data frame}

\item{sets}{a list of vectors.Each vector represets a table}
}
\value{
The function will reutrn a data frame.

The components of the data frame are RV coefficients between each pair of tables.
}
\description{
RV_table is the function of MFA package.
             The function aims to analyze the similarity between different table in a dataset
}

