% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{summary_eigen}
\alias{summary_eigen}
\title{summary_eigen}
\usage{
summary_eigen(x)
}
\arguments{
\item{x}{a mfa object}
}
\value{
The function will reutrn a data frame including 5 componenets.

1.singular value:the square root of eigenvalue.

2.eigenvalue: one of the properties contained in the mfa project.

3.cumulative1: the cumlulative of the eigenvalues in decreasing series.

4.Inertia: the proportion of each eigenvalue to the sum of all eigenvalues.

5.cumulative2: the cumlulative of the intertia in decreasing series.
}
\description{
summary_eigen is the function of MFA package.
             The function aims to analyze the property of eigenvalues.
}

