% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa}
\alias{mfa}
\title{mfa}
\usage{
mfa(data, sets, ncomps = NULL, center = TRUE, scale = TRUE)
}
\arguments{
\item{data:}{The working example consists of a (fictitious) wine tasting experiment.}
}
\value{
The function will return an object of class "mfa" with eigenvalues, common factor scores matrix, partial factor scores matrix and factor loadings.
}
\description{
mfa is the main function of MFA package
}
\examples{
\dontrun{
data_raw<-read.csv("data/wines.csv")
data<-data_raw[,2:54]
pick1<-list(seq(1,6),seq(7,12),seq(13,18),seq(19,23),seq(24,29),seq(30,34),seq(35,38),seq(39,44),seq(45,49),seq(50,53))
review<-mfa(data,pick1,scale=1) }
}

