% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa}
\alias{mfa}
\title{mfa}
\usage{
mfa(data, sets, ncomps = NULL, center = TRUE, scale = TRUE)
}
\arguments{
\item{data:}{data set(matrix or data frame)}

\item{sets:}{list of vectors indicating the sets of the varaiables (column of data)}

\item{ncomps:}{integer indicating the dimension (the number of eigenvalue needed to consider)}

\item{center:}{center is simlar to the arugement in the 'scale' function}

\item{scale:}{either a logical value or a numeric vector of length equal to the number of active variables in the analysis}
}
\value{
The function will return an object of class "mfa" with eigenvalues, common factor scores matrix, partial factor scores matrix and factor loadings.
}
\description{
mfa is the main function of MFA package
}
\examples{
\dontrun{
data_raw<-read.csv("data/wines.csv")
data<-data_raw[,2:54]
pick1<-list(seq(1,6),seq(7,12),seq(13,18),seq(19,23),seq(24,29),seq(30,34),seq(35,38),seq(39,44),seq(45,49),seq(50,53))
review<-mfa(data,pick1)}
}

