% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa}
\alias{mfa}
\title{mfa}
\usage{
mfa(data, sets, ncomps = NULL, center = TRUE, scale = TRUE)
}
\arguments{
\item{data} {data that need to be analysed(matrix or data frame)}
\item{sets} {list of vectors indicating the sets of variables (i.e. the blocks)}
\item{ncomps} {integer indicating how many number of components (i.e. factors) are to be extracted}
\item{center} {either a logical value or a numeric vector of length equal to the number of active variables in the analysis}
\item{scale} {a logical value,a number or a numeric vector of length equal to the number of active variables in the analysis}
}
\value{
result:...
}
\description{
a brief description
}
\examples{
\dontrun{
an example}
}

